Revert "Cleanup and moving around code for better readability"
This reverts commit f608d7a4260834b37b098b7f5f9e43df22a36dce.
