--- src/tracker_channels/mod.rs
@@ -1,8 +1,0 @@
-pub mod torrents;
-pub mod updates;
-pub mod shadow;
-pub mod whitelist;
-pub mod blacklist;
-pub mod peers;
-pub mod keys;
-pub mod stats;
