+++ src/tracker_channels/shadow.rs
@@ -2,13 +2,12 @@ use std::collections::HashMap;
 use log::debug;
 use serde_json::{json, Value};
 use crate::common::InfoHash;
-use crate::tracker::TorrentTracker;
-use crate::tracker_channels::stats::StatsEvent;
+use crate::tracker::{StatsEvent, TorrentTracker};
 
 impl TorrentTracker {
     pub fn channel_shadow_init(&self)
     {
-        let (_channel_left, channel_right) = self.shadow_channel.clone();
+        let (channel_left, channel_right) = self.shadow_channel.clone();
         tokio::spawn(async move {
             let mut shadow: HashMap<InfoHash, i64> = HashMap::new();
 
@@ -118,7 +117,7 @@ impl TorrentTracker {
 
     pub async fn channel_shadow_request(&self, action: &str, data: Value) -> (Value, Value)
     {
-        let (channel_left, _channel_right) = self.shadow_channel.clone();
+        let (channel_left, channel_right) = self.shadow_channel.clone();
         // Build the data with a action and data separated.
         let request_data = json!({
             "action": action,
