+++ src/tracker_channels/updates.rs
@@ -2,13 +2,12 @@ use std::collections::HashMap;
 use log::debug;
 use serde_json::{json, Value};
 use crate::common::InfoHash;
-use crate::tracker::TorrentTracker;
-use crate::tracker_channels::stats::StatsEvent;
+use crate::tracker::{StatsEvent, TorrentTracker};
 
 impl TorrentTracker {
     pub fn channel_updates_init(&self)
     {
-        let (_channel_left, channel_right) = self.updates_channel.clone();
+        let (channel_left, channel_right) = self.updates_channel.clone();
         tokio::spawn(async move {
             let mut updates: HashMap<InfoHash, i64> = HashMap::new();
 
@@ -116,7 +115,7 @@ impl TorrentTracker {
 
     pub async fn channel_updates_request(&self, action: &str, data: Value) -> (Value, Value)
     {
-        let (channel_left, _channel_right) = self.updates_channel.clone();
+        let (channel_left, channel_right) = self.updates_channel.clone();
         // Build the data with a action and data separated.
         let request_data = json!({
             "action": action,
