+++ src/tracker_channels/whitelist.rs
@@ -2,13 +2,12 @@ use std::collections::HashMap;
 use log::{debug, info};
 use serde_json::{json, Value};
 use crate::common::InfoHash;
-use crate::tracker::TorrentTracker;
-use crate::tracker_channels::stats::StatsEvent;
+use crate::tracker::{StatsEvent, TorrentTracker};
 
 impl TorrentTracker {
     pub fn channel_whitelist_init(&self)
     {
-        let (_channel_left, channel_right) = self.whitelist_channel.clone();
+        let (channel_left, channel_right) = self.whitelist_channel.clone();
         tokio::spawn(async move {
             let mut whitelist: HashMap<InfoHash, i64> = HashMap::new();
 
@@ -117,7 +116,7 @@ impl TorrentTracker {
     /* === Channel: Whitelist === */
     pub async fn channel_whitelist_request(&self, action: &str, data: Value) -> (Value, Value)
     {
-        let (channel_left, _channel_right) = self.whitelist_channel.clone();
+        let (channel_left, channel_right) = self.whitelist_channel.clone();
         // Build the data with a action and data separated.
         let request_data = json!({
             "action": action,
