+++ src/udp_service.rs
@@ -10,9 +10,7 @@ use tokio::time::timeout;
 
 use crate::common::{AnnounceEvent, AnnounceQueryRequest, InfoHash, maintenance_mode, PeerId};
 use crate::handlers::handle_announce;
-use crate::tracker::TorrentTracker;
-use crate::tracker_channels::stats::StatsEvent;
-use crate::tracker_channels::torrents::{TorrentEntry, TorrentEntryItem};
+use crate::tracker::{StatsEvent, TorrentEntry, TorrentEntryItem, TorrentTracker};
 use crate::udp_common;
 use crate::udp_common::{AnnounceInterval, AnnounceRequest, AnnounceResponse, ConnectRequest, ConnectResponse, ErrorResponse, get_connection_id, NumberOfDownloads, NumberOfPeers, Port, Request, Response, ResponsePeer, ScrapeRequest, ScrapeResponse, ServerError, TorrentScrapeStatistics, TransactionId};
